+++ ikii-back-commons/src/main/java/mx/ikii/commons/payload/request/order/OrderStatusRequest.java
@@ -1,0 +1,13 @@
+package mx.ikii.commons.payload.request.order;
+
+import lombok.Data;
+import mx.ikii.commons.domain.OrderSubStatus;
+
+@Data
+public class OrderStatusRequest {
+
+	private OrderSubStatus subStatus;
+
+	private String description;
+
+}